Updated Skip List to work a bit better. It still doesn't perform well on sizes>100,000 but much better. Also so code clean-up in other classes.
git-svn-id: https://java-algorithms-implementation.googlecode.com/svn/trunk@421 032fbc0f-8cab-eb90-e552-f08422b9a96a

+++ src/com/jwetherell/algorithms/numbers/Integers.java
@@ -47,7 +47,7 @@ public class Integers {
         StringBuilder builder = new StringBuilder();
         double temp = 0d;
         while (integer > 0) {
-            temp = ((double) integer) / 2d;
+            temp = integer/2d;
             integer = (int) temp;
             builder.append((temp > integer) ? 1 : 0);
         }
